Add files via upload
The DOM is a programming interface for web documents. It represents the page so that programs (like JavaScript) can change the document structure, style, and content. The DOM represents the document as nodes and objects; that way, programming languages can interact with the page.In this we will learn from basic to advance.